more refactoring and create compadre benchmark test
